Address PR comments - Remove TODO - Refactor duplicate code into dg-common
