+++ packages/datagateway-dataview/src/App.test.tsx
@@ -1,22 +1,9 @@
 import * as React from 'react';
-import App, { QueryClientSettingUpdater } from './App';
+import App from './App';
 import log from 'loglevel';
-import {
-  render,
-  screen,
-  waitFor,
-  type RenderResult,
-} from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import PageContainer from './page/pageContainer.component';
 import { configureApp, settingsLoaded } from './state/actions';
-import { StateType } from './state/app.types';
-import { Provider } from 'react-redux';
-import configureStore from 'redux-mock-store';
-import thunk from 'redux-thunk';
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { dGCommonInitialState } from 'datagateway-common';
-import { initialState as dgDataViewInitialState } from './state/reducers/dgdataview.reducer';
-import { createLocation } from 'history';
 
 jest
   .mock('loglevel')
@@ -90,56 +77,5 @@ describe('App', () => {
 
     // check that fallback UI is shown
     expect(await screen.findByText('app.error')).toBeInTheDocument();
-  });
-});
-
-describe('QueryClientSettingUpdater', () => {
-  const initialState: StateType = {
-    dgcommon: dGCommonInitialState,
-    dgdataview: dgDataViewInitialState,
-    router: {
-      action: 'POP',
-      location: { ...createLocation('/'), query: {} },
-    },
-  };
-  const renderComponent = (
-    state: StateType = initialState,
-    queryClient = new QueryClient()
-  ): RenderResult => {
-    const mockStore = configureStore([thunk]);
-    function Wrapper({
-      children,
-    }: React.PropsWithChildren<unknown>): JSX.Element {
-      return (
-        <Provider store={mockStore(state)}>
-          <QueryClientProvider client={queryClient}>
-            {children}
-          </QueryClientProvider>
-        </Provider>
-      );
-    }
-    return render(<QueryClientSettingUpdater queryClient={queryClient} />, {
-      wrapper: Wrapper,
-    });
-  };
-
-  beforeEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  it('syncs retry setting to query client when it updates', async () => {
-    const queryClient = new QueryClient({
-      // set random other option to check it doesn't get overridden
-      defaultOptions: { queries: { staleTime: 300000 } },
-    });
-    const { rerender } = renderComponent(initialState, queryClient);
-
-    initialState.dgcommon.queryRetries = 0;
-
-    rerender(<QueryClientSettingUpdater queryClient={queryClient} />);
-
-    expect(queryClient.getDefaultOptions()).toEqual({
-      queries: { staleTime: 300000, retry: 0 },
-    });
   });
 });
