+++ packages/datagateway-dataview/src/App.tsx
@@ -7,6 +7,7 @@ import {
   Preloader,
   BroadcastSignOutType,
   RequestPluginRerenderType,
+  QueryClientSettingsUpdaterRedux,
 } from 'datagateway-common';
 import {
   createBrowserHistory,
@@ -17,7 +18,7 @@ import {
 import log from 'loglevel';
 import React from 'react';
 import { Translation } from 'react-i18next';
-import { batch, connect, Provider, useSelector } from 'react-redux';
+import { batch, connect, Provider } from 'react-redux';
 import { AnyAction, applyMiddleware, compose, createStore, Store } from 'redux';
 import { createLogger } from 'redux-logger';
 import thunk, { ThunkDispatch } from 'redux-thunk';
@@ -106,27 +107,6 @@ const queryClient = new QueryClient({
   },
 });
 
-export const QueryClientSettingUpdater: React.FC<{
-  queryClient: QueryClient;
-}> = (props) => {
-  const { queryClient } = props;
-  const queryRetries = useSelector(
-    (state: StateType) => state.dgcommon.queryRetries
-  );
-
-  React.useEffect(() => {
-    if (typeof queryRetries !== 'undefined') {
-      const opts = queryClient.getDefaultOptions();
-      queryClient.setDefaultOptions({
-        ...opts,
-        queries: { ...opts.queries, retry: queryRetries },
-      });
-    }
-  }, [queryClient, queryRetries]);
-
-  return null;
-};
-
 document.addEventListener(MicroFrontendId, (e) => {
   const action = (e as CustomEvent).detail;
   if (action.type === BroadcastSignOutType) {
@@ -205,7 +185,7 @@ class App extends React.Component<unknown, { hasError: boolean }> {
           <Provider store={this.store}>
             <ConnectedRouter history={history}>
               <QueryClientProvider client={queryClient}>
-                <QueryClientSettingUpdater queryClient={queryClient} />
+                <QueryClientSettingsUpdaterRedux queryClient={queryClient} />
                 <DGThemeProvider>
                   <ConnectedPreloader>
                     <React.Suspense
